backend/pyproject.toml
[project]
name = "p1"
version = "0.1.0"
description = "Data Ingestion Tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.6.0",
    "django-environ>=0.13.0",
    "djangorestframework>=3.15.0",
    "langextract>=1.1.1",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pyproject-flake8>=7.0.0",
    "requests>=2.32.5",
]

[tool.flake8]
max-line-length = 120
